refactor: regroup extensions
