+++ ios/extensions/Date.swift
@@ -1,0 +1,15 @@
+//
+//  Extensions.swift
+//  KeyboardController
+//
+//  Created by Kiryl Ziusko on 04/07/2024.
+//  Copyright © 2024 Facebook. All rights reserved.
+//
+
+import Foundation
+
+public extension Date {
+  static var currentTimeStamp: Int64 {
+    return Int64(Date().timeIntervalSince1970 * 1000)
+  }
+}
