+++ ios/extensions/UIResponder.swift
@@ -2,32 +2,12 @@
 //  Extensions.swift
 //  KeyboardController
 //
-//  Created by Kiryl Ziusko on 10/06/2023.
-//  Copyright © 2023 Facebook. All rights reserved.
+//  Created by Kiryl Ziusko on 04/07/2024.
+//  Copyright © 2024 Facebook. All rights reserved.
 //
 
 import Foundation
 import UIKit
-
-public extension CGFloat {
-  static func interpolate(inputRange: [CGFloat], outputRange: [CGFloat], currentValue: CGFloat) -> CGFloat {
-    let inputMin = inputRange.min() ?? 0
-    let inputMax = inputRange.max() ?? 1
-    let outputMin = outputRange.min() ?? 0
-    let outputMax = outputRange.max() ?? 1
-
-    let normalizedValue = (currentValue - inputMin) / (inputMax - inputMin)
-    let interpolatedValue = outputMin + (outputMax - outputMin) * normalizedValue
-
-    return interpolatedValue
-  }
-}
-
-public extension Date {
-  static var currentTimeStamp: Int64 {
-    return Int64(Date().timeIntervalSince1970 * 1000)
-  }
-}
 
 public extension UIResponder {
   private weak static var _currentFirstResponder: UIResponder?
@@ -49,16 +29,6 @@ public extension Optional where Wrapped == UIResponder {
       return ((self as? UIView)?.superview?.tag ?? -1) as NSNumber
     #else
       return (self as? UIView)?.superview?.reactTag ?? -1
-    #endif
-  }
-}
-
-public extension UIScrollView {
-  var reactViewTag: NSNumber {
-    #if KEYBOARD_CONTROLLER_NEW_ARCH_ENABLED
-      return (superview?.tag ?? -1) as NSNumber
-    #else
-      return superview?.reactTag ?? -1
     #endif
   }
 }
@@ -85,23 +55,5 @@ public extension Optional where Wrapped: UIResponder {
 
     // UIScrollView is not found
     return -1
-  }
-}
-
-public extension UIView {
-  var globalFrame: CGRect? {
-    let rootView = UIApplication.shared.keyWindow?.rootViewController?.view
-    return superview?.convert(frame, to: rootView)
-  }
-}
-
-public extension UITextInput {
-  var canSelectionFitIntoLayout: Bool {
-    guard let selectedRange = selectedTextRange else { return false }
-
-    guard let range = textRange(from: selectedRange.start, to: selectedRange.end) else { return false }
-    let rect = firstRect(for: range)
-
-    return rect.origin.x.isFinite && rect.origin.y.isFinite
   }
 }
